Move `AbstractObject` and `AbstractTensor` into `abstract` sub-package
