+++ syft/execution/placeholder.py
@@ -1,6 +1,6 @@
 import syft
 from syft.generic.frameworks.hook import hook_args
-from syft.generic.tensor import AbstractTensor
+from syft.generic.abstract.tensor import AbstractTensor
 from syft.workers.abstract import AbstractWorker
 from syft_proto.execution.v1.placeholder_pb2 import Placeholder as PlaceholderPB
