+++ syft/frameworks/torch/tensors/decorators/logging.py
@@ -1,5 +1,5 @@
 import syft
-from syft.generic.tensor import AbstractTensor
+from syft.generic.abstract.tensor import AbstractTensor
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.overload import overloaded
 from syft.workers.abstract import AbstractWorker
