+++ syft/frameworks/torch/tensors/interpreters/additive_shared.py
@@ -9,7 +9,7 @@ from syft.frameworks.torch.mpc import securenn
 from syft.frameworks.torch.mpc import fss
 from syft.generic.utils import memorize
 
-from syft.generic.tensor import AbstractTensor
+from syft.generic.abstract.tensor import AbstractTensor
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.overload import overloaded
 from syft.workers.abstract import AbstractWorker
