+++ syft/frameworks/torch/tensors/interpreters/autograd.py
@@ -1,7 +1,7 @@
 import torch
 
 import syft
-from syft.generic.tensor import AbstractTensor
+from syft.generic.abstract.tensor import AbstractTensor
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.overload import overloaded
 from syft.workers.abstract import AbstractWorker
