+++ syft/frameworks/torch/tensors/interpreters/private.py
@@ -8,7 +8,7 @@ from syft.generic.frameworks.overload import overloaded
 
 
 from syft.workers.abstract import AbstractWorker
-from syft.generic.tensor import AbstractTensor
+from syft.generic.abstract.tensor import AbstractTensor
 
 
 class PrivateTensor(AbstractTensor):
