+++ syft/generic/abstract/object.py
