+++ syft/generic/pointers/object_pointer.py
@@ -11,7 +11,7 @@ from syft.generic.frameworks.hook.hook_args import register_forward_func
 from syft.generic.frameworks.hook.hook_args import register_backward_func
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.types import FrameworkTensor
-from syft.generic.object import AbstractObject
+from syft.generic.abstract.object import AbstractObject
 from syft.messaging.message import ForceObjectDeleteMessage
 from syft.workers.abstract import AbstractWorker
